Refactor: rename packages and fixtures from esdb to kurrentdb
