+++ docs/documentation/code/recipes/integrate/kurrent_01.py
@@ -1,7 +1,7 @@
 from esdbclient import EventStoreDBClient
 
 from event_sourcery.event_store import StreamId
-from event_sourcery_esdb import ESDBBackendFactory
+from event_sourcery_kurrentdb import ESDBBackendFactory
 
 client = EventStoreDBClient(uri="esdb://localhost:2113?Tls=false")
 factory = ESDBBackendFactory(client)
