+++ docs/documentation/code/test_recipes.py
@@ -59,10 +59,10 @@ def test_sqlalchemy_backend(base_with_configured_es_models):
     assert backend.event_store.position == 0
 
 
-def test_eventstore_db():
+def test_kurrent_db():
     pass
     # from esdbclient import EventStoreDBClient
-    # from event_sourcery_esdb import ESDBBackendFactory
+    # from event_sourcery_kurrentdb import ESDBBackendFactory
     #
     # client = EventStoreDBClient(uri="esdb://localhost:2113?Tls=false")
     # factory = ESDBBackendFactory(client)
