+++ event_sourcery_kurrentdb/stream.py
