+++ event_sourcery_kurrentdb/subscription.py
@@ -9,7 +9,7 @@ from esdbclient import EventStoreDBClient, RecordedEvent
 
 from event_sourcery.event_store import Position, RecordedRaw
 from event_sourcery.event_store.interfaces import SubscriptionStrategy
-from event_sourcery_esdb import dto
+from event_sourcery_kurrentdb import dto
 
 
 class BuilderCallable(Protocol):
