+++ pyproject.toml
@@ -10,7 +10,7 @@ license = "MIT"
 packages = [
     { include = "event_sourcery" },
     { include = "event_sourcery_django" },
-    { include = "event_sourcery_esdb" },
+    { include = "event_sourcery_kurrentdb" },
     { include = "event_sourcery_sqlalchemy" },
 ]
 
@@ -39,7 +39,7 @@ deepdiff = "*"
 
 [tool.poetry.extras]
 sqlalchemy = ["SQLAlchemy"]
-esdb = ["esdbclient"]
+kurrentdb = ["esdbclient"]
 django = ["django"]
 
 [tool.pytest]
@@ -57,7 +57,7 @@ branch = true
 source = [
     "event_sourcery",
     "event_sourcery_django",
-    "event_sourcery_esdb",
+    "event_sourcery_kurrentdb",
     "event_sourcery_sqlalchemy",
 ]
 omit = [
