+++ tests/event_store/subscriptions/test_in_transaction_subscription.py
@@ -7,7 +7,7 @@ from tests.factories import AnEvent, OtherEvent, an_event
 from tests.matchers import any_record
 
 pytestmark = pytest.mark.skip_backend(
-    backend="esdb", reason="ESDB don't have transactions"
+    backend="kurrentdb", reason="KurrentDB don't have transactions"
 )
